fix: client component translations
